requirements.txt
requests~=2.27.1
discord~=1.7.3
python-dotenv~=0.20.0